misc: vpumgr: fix on FuRong's review comment
* use offsetof
* use dev_xxx instead of pr_xxx
* change vcm msg payload size to fit page size
* add function description to few important functions
* add checks on in_len in vcm_submit()
* add idr_destroy() in error of vcm_init
* changed smm_xxx ioctl handler return type as int
* add check for not found case in smm_unimport

+++ drivers/misc/vpumgr/vpu_smm.h
@@ -20,10 +20,10 @@ int smm_fini(struct vpumgr_device *vdev);
 int smm_open(struct vpumgr_smm *sess, struct vpumgr_device *vdev);
 int smm_close(struct vpumgr_smm *sess);
 
-long smm_alloc(struct vpumgr_smm *sess, struct vpumgr_args_alloc *arg);
-long smm_import(struct vpumgr_smm *sess, struct vpumgr_args_import *arg);
-long smm_unimport(struct vpumgr_smm *sess, int *p_dmabuf_fd);
-long smm_ptr2vpu(struct vpumgr_smm *sess, unsigned long *arg);
+int smm_alloc(struct vpumgr_smm *sess, struct vpumgr_args_alloc *arg);
+int smm_import(struct vpumgr_smm *sess, struct vpumgr_args_import *arg);
+int smm_unimport(struct vpumgr_smm *sess, int *p_dmabuf_fd);
+int smm_ptr2vpu(struct vpumgr_smm *sess, unsigned long *arg);
 
 int smm_debugfs_stats_show(struct seq_file *file, struct vpumgr_smm *sess);
